requirements.txt
pytesseract
Pillow